refactor: move cli, update_manager, uninstall_manager into cli/ package
