+++ src/aish/cli/uninstall_manager.py
